refactor: change type of att TotalPrice

+++ server/Models/OrderItems.cs
@@ -2,7 +2,7 @@ public class OrderItem
 {
     public int Id { get; set; }
     public DateTime CreatedAt { get; set; }
-    public double TotalPrice { get; set; }
+    public float TotalPrice { get; set; }
     public int ShoppingCartItemId { get; set; }
 }
